+++ telematic_system/telematic_apps/web_app/server/utils/verify_token.js
@@ -12,7 +12,7 @@
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
  * License for the specific language governing permissions and limitations under
  * the License.
- * 
+ *
  * Description:
  * verifyToken function to decode authentication token passed by user request and return the user information.
  */
